+++ src/middleware/query.ts
@@ -1,17 +1,13 @@
 import * as http from 'http'
 import * as parseUrl from 'parseurl'
 import * as qs from 'qs'
+import * as Fox from '../index'
 
-const query = (options: Object | Function = {}) => {
-  let queryparse = qs.parse
-
-  if (Function.isInstance(options)) {
-    queryparse = <(str: string, options?: qs.IParseOptions) => any>options
-    options = {}
-  }
+const query = (app: Fox) => {
+  let queryparse: Function = app.get('query.parser') || qs.parse
 
   return (req: http.IncomingMessage, res: http.ServerResponse, next: () => void) => {
-    req.query = req.query || queryparse((parseUrl(req) as any).query, <Object>options)
+    req.query = req.query || queryparse((parseUrl(req) as any).query, {})
 
     next()
   }