+++ src/routing/Route.ts
@@ -1,5 +1,4 @@
 import * as http from 'http'
-import * as pathToRegExp from 'path-to-regexp'
 import { Encapsulation } from '../exeptions'
 import httpMethods from './httpMethods'
 
@@ -13,7 +12,7 @@ declare module Route {
   }
 
   export interface RouteObject {
-    path: RegExp,
+    path: string | RegExp,
     controller: Encapsulation
   }
 }
@@ -22,18 +21,11 @@ declare interface Route {
   [key: string]: any
 }
 
-/**
- *
- */
 class Route {
   public routes: Route.Routes = {}
 
   protected _prefix: string
 
-  /**
-   *
-   * @param {String?} prefix
-   */
   constructor(prefix: string = '') {
     this._prefix = prefix
 
@@ -44,41 +36,27 @@ class Route {
     })
   }
 
-  /**
-   *
-   * @param {!String} method
-   * @param {!String} path
-   * @param {!Function} controller
-   * @private
-   */
   protected _push(method: string, path: string, controller: Route.Controller) {
     path = `${this._prefix}${path}`.replace(/\/$/, '')
 
-    let _path = pathToRegExp(path, [], { sensitive: true })
-
     this.routes[method].push({
-      path: _path,
+      path,
       controller: new Encapsulation((req, res, next: () => void, ...args: Array<any>) => controller(req, res, next, ...args))
     })
+
+    return this
   }
 
-  /**
-   *
-   * @param {!String} path
-   * @param {!Function} controller
-   */
   any(path: string, controller: Route.Controller) {
     httpMethods.map((method) => this._push(method, path, controller))
+
+    return this
   }
 
-  /**
-   *
-   * @param {!(String[])} methods
-   * @param {!String} path
-   * @param {!Function} controller
-   */
   oneOf(methods: Array<string>, path: string, controller: Route.Controller) {
     methods.map((method) => this._push(method.toUpperCase(), path, controller))
+
+    return this
   }
 
   /**
@@ -102,6 +80,8 @@ class Route {
 
       this.any(_path, <Route.Controller>_middleware)
     }
+
+    return this
   }
 }