+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 
 
+## [v0.5.0](https://github.com/foxifyjs/foxify/releases/tag/v0.5.0) *(2018-03-08)*
+
+**Implemented enhancements:**
+- added clustering
+- added multiple model relations [`hasOne`, `hasMany`]
+- added multiple options [`x-powered-by`, `routing.strict`, `routing.sensitive`, `json.escape`]
+- added multiple settings [`env`, `url`, `port`, `clusters`, `json.replacer`, `json.spaces`, `query.parser`]
+- improved usage
+
+
 ## [v0.4.0](https://github.com/foxifyjs/foxify/releases/tag/v0.4.0) *(2018-02-22)*
 
 **Implemented enhancements:**